add help module

+++ utils/messages.py
@@ -14,7 +14,8 @@ async def clear_messages(bot, messages):
     """Delete messages after a delay"""
     await asyncio.sleep(constants.SPAM_DELAY)
     for message in messages:
-        await bot.delete_message(message)
+        if not message.channel.is_private:
+            await bot.delete_message(message)
 
 
 class MessageManager:
